Remove int/uint from examples
int/uint has been renamed to isize/usize but it is not really appropriate for
use in these examples.

Where int/uint cannot be constrained they are replaced w/ i64 and u64 respectively.
Otherwise we simply remove the annotations and let rustc constrain them to the
expected type.

+++ examples/zguide/weather-client/main.rs
@@ -8,7 +8,7 @@
 
 extern crate zmq;
 
-fn atoi(s: &str) -> int {
+fn atoi(s: &str) -> i64 {
     s.parse().unwrap()
 }
 
@@ -25,9 +25,9 @@ fn main() {
 
     let mut total_temp = 0;
 
-    for _ in range(0u, 100) {
+    for _ in range(0, 100) {
         let string = subscriber.recv_string(0).unwrap().unwrap();
-        let chks: Vec<int> = string.as_slice().split(' ').map(|x| atoi(x)).collect();
+        let chks: Vec<i64> = string.as_slice().split(' ').map(|x| atoi(x)).collect();
         let (_zipcode, temperature, _relhumidity) = (chks[0], chks[1], chks[2]);
         total_temp += temperature;
     }
